refactorizacion a usar objeto menu
