+++ menu.h
@@ -1,0 +1,12 @@
+#ifndef MENU_H
+#define MENU_H
+
+#include <string>
+using namespace std;
+
+int menuPrincipal(int& opcion);
+void limpiarPantalla();
+void confirmarVolverAlMenu();
+char volverAlMenu(char respuesta_si_no);
+
+#endif
