+++ validacion.h
@@ -10,12 +10,8 @@ void noLetras(string& valor);
 void noCaracteresEspeciales(string& valor);
 void noRepetidos(string& valor);
 int dentroDelRango(int& valor, int min, int max);
-char volverAlMenu(char respuesta_si_no);
 bool validarCedula(string cedula);
 bool validarEdad(int edad);
 bool validarNota(float nota);
-int menuPrincipal(int& opcion_seleccionada);
-void limpiarPantalla();
-void confirmarVolverAlMenu();
 
 #endif
